ui/app/RewardsPage,RewardsContainer: Validatory Subsidy, DRY
